Make keyboard-driven menu popups appear by the item providing the popup instead of the arbitrarily located mouse cursor. This includes all RCMTreeViews in qltk and browsers via RCMTreeView#popup_menu as well as the play and timer buttons in qltk/controls.py via qltk.popup_menu_under_widget. The position is calculated correctly for LTR and RTL environments.
Open concern: the undocumented position function returns a 3-tuple of (x, y, ?) where ? must be an integer, but seems to have no effect. What value should it have? My best guess is it may be a screen index, but I cannot test it easily with my single screen.

Also fix an exception in edittags.py and playlists.py when no rows are selected and menu-key is pressed, and an exception in filesel.py when a directory is removed out from under dircache at an inopportune time.