Convert file extensions to lowercase on move. Closes issue #66.
Rather than make a larger change to add a configuration option for
lowercasing of extensions, this patch tries to do sensible defaults.
If the pattern result includes an extension which is case-
insensitive equal to the original extension, the result's extension
is used, but if it doesn't include an extension, the extension is
automatically lowercased.