gstbe: Do not pause network streams, destroy pipeline for them
From bb51786bf414c14b9831de6fbb1c37ccb7bfdb52 Mon Sep 17 00:00:00 2001
Do the same thing the Xine backend does:  If the request is to pause,
check if song is a file.  Pause if file, stop (in GStreamer backend
meaning destroy pipeline) if it is not.  On unpause, the code to
initialize a new pipeline if it does not exist is already in place.

Without this, radio streams that were paused for a long time will play
what is left in the buffer and then fail, since no matching followup
data can be received.
---
 quodlibet/quodlibet/player/gstbe.py |    6 ++++--
 1 files changed, 4 insertions(+), 2 deletions(-)