More language/gettext fixes: Make LANGUAGE work when it includes en/en_XX; Convert osx locale identifiers;
gettext fails in case LANGUAGE contains en/en_XX identifiers as it doesn't know
that the msgid's could be used as a fallback. QL could include a dummy en.po but
dependencies like gtk+ would need to dos so as well. Instead work around by appending
"C" to each en/en_XX variant which makes gettext disable translation if reached.
Undocumented, but seems to work.

For osx locale identifier the language format is BCP 47 as well,
so convert that to ISO 639-1 if possible.