Stop player before disabling plugins. Closes #1567
If at the time plugins get disabled the player is still active
(an active GStreamer pipeline exists) the plugin handler
will rebuild the pipeline for each disabled plugin. This leads
to noise as the rebuild pauses playback shorty and in case
a stream is playing will block for a short time due to
setting up a new connection.