Makes album-heuristics consider date-tag
This is a simple heuristic that ensures that albums with the same name
are considered distinct if they have different date-tags.

As by their nature albums have an unambigious release date, there should
be nothing messed by this.

This solves e.g., different Albums named "In Dub" to be merged into one.
This should also solve the initial behaviour described in #559. Though
this patch also allows for same-named albums to be in the same directory.