gstbe: Implement direct sink volume control (e.g. for pulsesink, directsoundsink). Closes #1389. See #1512.
Instead of taking the replaygain adjustment and the user volume adjustment
and applying it using an extra volume element, apply the user volume
adjustment directly to the audio sink.

This keeps the QL volume widget in sync with the pulseaudio application
volume level and removes the delay when changing the volume from QL
(especially noticable using directsoundsink under Windows).

Because this makes the volume commonly displayed in QL lower as before
and to behave more like other players and the Gnome sound settings,
use a cubic volume scale for the main volume widget.

Remaining problems are caused by the fact that get/setting the volume
is only possible if QL is paused/playing and that pulsesink
doesn't notify of volume changes when paused or when creating
a pulseaudio connection. To improve this QL will update the volume
on pipeline state changes and on any user events on the main volume
widget.