config: rework default value handling
* Config now has a defaults attribute which is itself a Config
providing values in case a parsing error happens or
the value is missing. This allows to set default values
using the setxxx methods.

* In case of raising on a parsing error, all options with
an explicitly passed default or a set default will return
the default instead of raising. This is usually what one wants
anyway and reduces the risk of QL breaking in case of wrong values
in the config files.