Fix border drawing with CSD/wayland.
The hack assumes that the window widget has the same allocation
than the child. This is not the case with CSD so take the child
allocation instead.

Also remove the top borders for exfalso; this works better
now with the newer gtk theme as the page labels have enough
padding.