player: add a "seekable" property and implement it in the GStreamer backend
Allows us to disable the seek bar in case the current song isn't seekable.
Also use it to decide if we want to rebuffer on unpause instead of
using the duration query error case (which might be unreliable depending on the
pipeline state)

All other backends default to True and in case there is no song to False.