tagedit: fix removing non-shared tag values from multiple songs
In case multiple songs were selected and a tag removed for which
the values were different for each song (e.g. musicbrainz_trackid)
it would only remove the value from one song.

This bug got exposed when AudioFile.remove was changed to not
remove all values in case the value wasn't found (that might explain
why it did that in the first place)