wayland: fix window size restore with gtk+ trunk
gtk+ now tries to compensate for the CSD when handling window sizes.
This includes get_size() and resize() but not the configure event.
Call get_size() instead of using the event values which contain CSD.

This also affects CSD windows on X11, but we don't have any which restore their state.