Remove cddb plugin; it's broken
I can't get a single album to match...
The generation of the disc id depends on the disk layout which we don't have (and just guess based on the song lengths).
freedb even returns an error if I try to match the disk offsets to a known release so maybe the matching
got stricter somehow?

Switching to the musicbrainz freedb gateway works but gives bad results.

So just remove it.. if someone wants to bring it back, feel free to go ahead.