Verify trash folder, fall back to home trash
Consolidate trash folder finding and building into a single function
that makes the trash dir and its required subdirs if needed. This routine
also verifies that writes to the trash directory structure are possible
before returning it, and will fall back correctly to the home trash
folder if it has to.

Ref http://standards.freedesktop.org/trash-spec/trashspec-latest.html

Fixes quodlibet/quodlibet#1339, or at least makes the exception go away
for many important cases.