move highlight drawing to Gtk.CellArea; allow passing properties to TreeViewColumn
TreeViewColumn overrides don't allow setting properties so bypass them and clean up the
interface.

Instead of drawing the highlight in the cell renderers, draw them one level up
in the CellArea. This results in less draws and fixes leftover drawing artefacts
in some cases as apparently the cell renderer is only supposed
to draw in the cell area..