search: make "ae" match "æ"
Support characters which can be decomposed into multiple charaters.
This only works if the characters are following each other and there
are is no overlapping of possible decomposed groups.

e.g. "a[e]" wont match "æ" and "LLL" only matches "ỺL" but not "LỺ"

but for the most basic things it should work.