tests: switch to pytest as the main test runner
Allows us to get rid of a bunch of test discovery code
and gives us a few new features.

"--to-run=TFoo,TBar" still works by passing the class
names to pytest using the "-k" switch. This accepts
more stuff, but for the basic class name list it
should give the same results.

To skip network and code quality tests pytest markers
are used:
  @pytest.mark.quality
  @pytest.mark.network