Remove get_fs_encoding, no longer used
While it's nice that we can use the same encoding as glib,
in reality there is not much difference. Windows is utf-16,
osx is utf-8 and on linux every desktop will have utf-8 by default.

Also with Python 3 we can't use a different encoding as the
one provided by Python itself, otherwise we can't encode
str paths with surrogates there.