Added ModeBar
The shuffle and repeat buttons are subtly different in appearance when depressed and when not. Further, their meaning is not clear until the hover hint is displayed.  I added a small ModeBar to provide clear visual feedback on the state and meaning of these two buttons when used.

This is a very first sally on my part with QL mods and a pull request. Am open to feedback on style and implementation.