gtk: add some css overrides for Adwaita to make the entry/button height depend on the font size
This is tracked upstream in https://bugzilla.gnome.org/show_bug.cgi?id=761435
but didn't make it into 3.22, so we have to live with it for a while.

In cases where the font size is small (Windows/OSX/Gnome with reduced size)
the entries/buttons are too large. Add some css to make it depend
on the font size instead.

This adds a theme overrider class which allows to apply css only when
a specific theme is active.