pickle: switch to protocol v2
The problem with 2 was that it calls __setitem__ which we override
and which is (1) slow and (2) we are stricter now with what can be
assigned than be have before, so things can fail at load time.

To work aorund that we use the new class lookup function and create
a temporary class for each real class which just inherits from dict
and doesn't have a custom __setitem__. After pickle is done
we replace the __class__ for each item with the real AudioFile
subclass.

This now allows us to get rid of the __setitem__ safety checks in
AudioFile and is faster as well (50 ms here, even more on py3)