Rework parts of the file/folderchooser handling
Newer GTK+ added an interface for opening the native file chooser on
Windows so add a simple interface using the new/old interface depending
on the gtk+ version.

Replaces some simple filechooser usage which didn't need API not available
on Windows. One thing removed was the "watch for new files" checkbox, which
was a bit confusing anyway imo.

Some places saved/restored the last visible directory for the file chooser.
This replaces it with a global path saved in the config.

TODO: The remaining, less simple file chooser cases need to be replaced with
the chooser and an addition dialog for further actions.