Improve waveform seekbar hoover time indication
Before this patch only a tiny line was drawn to indicate the current position.
With this patch the hover color is used to display the current playing
position if the mouse hovers before this position and the elapsed color is
used to display the new position, if the mouse hovers after the position the
elapsed color is used to display the current playing position and the hover
color is used to display the new position. This might sound confusing but
works intuitively very well.