Some initial mypy support. See #2299
Adds the basic configuration and a (skipped) test in the test suite.
Also rework some of the code to make mypy happy without adding
any annotations for now.

You can run it by executing "mypy ." in the directory setup.py is in.

There are still 136 errors reported which need to be fixed.

mypy is quite slow and takes 18sec here for the whole codebase and
14sec with the experimental "incremental" mode.