Fix randomalbum plugin getting active when activating the browser. Fixes #2726
While setting the song to None and then start playing gives the same result as
reset() we emit one more signal, which confuses the random album plugin as it thinks
that the song list has ended and starts scheduling a new album.

Go back to the old method instead for now.