Fix waveform calculation
This fixes an issue with the calculation of the waveform that became
apparent with the higher placeholder waveform. If a song was paused and
the player was closed and then reopened again the placeholder looked
"funny". The culprit for that was that the widget of the
WaveformSeekBar was not yet realized and thus not enough values were
produced. This patch fixes that issue by checking if the widget was
realized before starting the pipeline.

The performance should also be improved a little bit if the window is
not visible as we check for that before updating anything.