translations: try to load translations for all locale dirs before giving up. See #2819
The current approach was to search for the first existing locale dir
and try to load translations from there, but the translations could be
in multiple dirs and the first one might be the wrong one.

Instead iterate through all of them and try to load the translations
before giving up.