fifo: when reading a response times out assume the other instance is broken. Fixes #2754
When the main instance deadlocks (due to GStreamer bugs for example) we can still
write to the fifo (I assume until the buffer is full) but reading will time out.
In case we time out just delete the fifo so we continue creating our own
and take over as the primary instance. Better to have two instances then
QL not working until a restart.

We currently time out after 10 seconds.