qltk/edittags: Allow saving tag if present in all songs but value differ
If a set of songs all had a tag defined but with differing values, the
edit tags dialog would not allow changing to one of the existing values.

For example, two albums with tags album='A' and album='A (disc 2)' could
not directly be set to album='A' together.

Fixes #2686