themeswitcher: also list themes which target specific gtk versions
It was only looking for gtk-3.0 directories but with gtk 3.18 or so
a features was added that a theme could be named "gtk-3.20" and any gtk
version equal or newer than that would use it.

Copy this logic in the theme switcher plugin. This for examples makes
the Breeze theme show up.