mmkeys/gnome: check that the dbus object supports the interface before using it. Fixes #3056
When mate-settins-daemon is installed but mate isn't run it still gets started (at least under Debian)
but it doesn't provide the mmkeys interface. This resulted in the mate backen dbeing selected and
in an error on start.

This extends the code to check for the interface instead for the dbus name and also handles dbus errors
by logging them instead of preventing QL from starting.