wininst: rework .pyc generation
Only use the old-style .pyc files for the stdlib and dependencies where we remove
the .py files and use __pycache__ directories for QL itself.
This is more in line with the default usage of .pyc files.

Also use the new unchecked-hash format so that we get reproducible .pyc files
and .py files aren't checked unless explicitely wanted (sentry etc)

All in all this makes startup a bit faster.