Rework and merge operon/quodlibet completions into one file
This allows to easily share helpers and variables that are used by both
operon() and quodlibet() completion functions.

-- Helpers and cache

We use helpers to load audio extensions dynamically, by calling into QL
code. This might break easily in the future though, so this is no
perfect solution, but better that hard-code a list of audio extensions
that would never be maintained

Additionally, note that we cache the result of these helpers, so that we
don't have to call Python each time the user hit the completion. That
would be poor user experience on slow computers.

-- Installation

When a script defines completions for more than one command, the
"traditional" way to install on UNIX platforms it is to add a symlink
for each command, so for example here we would install the file
`quodlibet.bash` to `/usr/share/bash-completion/completions/quodlibet`,
and then install a symlink `operon -> quodlibet`.

However Quodlibet works on various platforms, including Windows+MSYS2,
which includes bash and bash-completion, but where symlinks are a
complicated topic, and best avoided.

As a result, the current QL install code doesn't deal with symlinks.

To keep this behavior, we choose NOT to create a symlink `operon ->
quodlibet`, and instead we prefer to just install a copy of the file,
under the name operon.

Signed-off-by: Arnaud Rebillout <elboulangero@gmail.com>