Minor tweaks needed
Hello! After my apparently unnecessary cry for help in the previous update, I managed to get this thing working. It need some minor tweaks to make it run smoothly, but as it stands, it works. There's a couple of known issues that I might revisit at a later date, or somebody else might want to fix them. Thing is, they can be a bit annoying but nothing too drastic.
These are some known bugs:

 * Pausing and unpausing a song a lot of times can make the plugin crash
 * If you have a song that has ELRC syntax, for it to be detected you must play another song first.
 * The highlight color in the plugin's options requires a restart for it to take effect.
 * The last line of the lyrics never shows, a workaround I use is to create a new empty line at the bottom so it shows the last line and it doesn't show the empty one.

That's what I can think of right now, I don't think there's more issues. I will open the pull request for review. The plugin autodetects if your song has LRC or ELRC syntax and uses either one accordingly. It also supports embedded tags or external files. If a file is found, the tag is ignored, if there's no file, the tag is used.
I'll try to update a song with ELRC syntax so that you can check out how the plugin works.