Finished plugin
At last, the plugin is working as intended with all minor imperfections fixed. I tested all the scenarios that came to mind, and the plugin works as it should in each of them. It still needs to be restarted to change the color of the highlight, but that is just how gtk tags work, not really much I can do about it. Anyway, feel free to test it and provide feedback. I say it has no issues, but I would not be amazed if it actually does have bugs. After all, the code is kinda clunky, and things kept breaking for some reason and I had to go back to previous sections to fix it. It supports .lrc files, tags, and ELRC syntax. I also reduced the lines sizes to accomodate to the 79 columns limit that exists for some reason.