Support for mock-confirmer for ConfirmationPrompt (#3396)
Adds support for passing a Confirmer-class to PlaylistsBrowser. This
class will be instanced with the arguments to set up a gtkdialog and
afterwards its run-method will be called.

The result of the run-method is a Gtk.Responsetype, for example
Gtk.Responsetype.YES. The resulting responsetype is compared to the
RESPONSE_INVOKE-attribute of the class, and if the one returned from
the run-method is equal, it will be iterpreted as the user having
accepted the prompt.

This enables passing in either the standard ConfirmationPrompt, which
creates an actual Gtkdialog, or alternatively a mock class for e.g.
unit testing.

Added a pair of unit tests that make use of such a mock-confirmer for
removal of playlists, which serve as an example of its usage.

Also changed the playlist-removal prompt to use ConfirmationPrompt,
as a further example of usage.

ConfirmationPrompt now uses the enum Gtk.ResponseType.YES as opposed
to a hardcoded 1.