discord_status: Handle DiscordNotFound.
This seems to happen only on cold boot, when Quod Libet is launched and
tries to talk to Discord before Discord has had a chance to run
(and presumably create the IPC bridge).

This exception is not seen if Discord (and optionally, Quod Libet) is
quit and Quod Libet tries to talk to Discord.