Fix regression of --print-query
"quodlibet --print-query=asdf" results in a timeout if the music
collection is large and the main quodlibet instance is started with
the --debug flag because currently Query is unnecessarily instantiated
for every song.  This commit fixes this, which results in a
performance increase even without --debug.

I think commit e40270f70 introduced the regression.