Fix crash in console when autocompleting
Some objects (like custom classes) don't have a defined __repr__.
When an object like this is the default argument of a function,
we currently receive an uncaught TypeError exception.

Instead, use `repr` to get the string representation of the object.
When __repr__ isn't defined, it falls back to returning a
reasonable representaion of most obects.