Stop query date regex from matching > 4 digit year (#4162)
Some queries are ambiguous between dates and numbers, e.g. "2002" could
be a year or a number. In order to simplify quering parsing, Quod Libet
stores this ambiguous subqueries as the "NumexprNumberOrDate" type.

When parsing a query, Quod Libet will use this type if the string
matches a regex for parsing dates. However, before this commit, the
regex will match the first four digits of any four or more digit number.
This results in a parsing error for large numbers because of the
remaining digits after the four that get matched and parsed.

Instead, use a negative lookahead to limit the year part of the date to
four digits. Quod Libet will fall back to using an unambiguous number
representation for large numbers.

As an alternative, I tried to allow >4 digit years for future-proofing,
but a 4 digit year assumption is baked in too deeply, e.g. in Python's
`strptime` function.

Fixes #4161.