Don't insert blank tag when doing column sort magic
If a user has a column with a tag pattern, and tries to sort by this
column, Quod Libet will try to make sorting work intuitively by mangling
the pattern and removing or replacing tags.

For example, "~#track" tags in patterns will be removed, on the
assumption that the user will prefer the default ordering, which already
looks at track numbers.

This commit fixes an issue where this is done incorrectly, and a tag
like "<~#track>" is replaced with an empty tag "<>", producing an
exception when sorting by the column.

It also adds several missing tags so that magic sorting will also work
for them.

Fixes #4164.