tests: stop setting XDG_RUNTIME_DIR
this was added in 642ac16482dc500d6 and I'm not sure if this was just
to cover all possible xdg paths, or if I had some specific thing in mind there.

Anyway, in the context of flatpak this seems to be used by glib-networking to load TLS
certificates from the host, and depending on when that loading happens and when we set
that env var things will fail.

Just try not setting XDG_RUNTIME_DIR for the tests and use the host one.