Add disjunction / fallback tag expression
Add a new piece of tag pattern syntax, with the operator `||`.
This operator returns the tag pattern on the left side, if it
is not empty, and otherwise returns the tag pattern on the
right side.

It is implemented so as to be chainable, e.g. one can do

    <<composer>||<albumartist>||<artist>>

And the first non-empty expression will be used.

In documentation, this syntax will be called the "fallback" tag
expression, for greater clarity.