gdist: fix create_po command in case LANG isn't set
msginit defaults to LANG for the language it writes into the .po
file and fails if LANG is not sure, which is usually only the case
on Windows. And everwhere else it would write the wrong language
ID into the file.

Explicitely pass the locale to msginit to fix this.