Update OSX build scripts and associated data files.
Update the build scripts and associated data used to build QL for
the Mac platform.  The updated scripts, when used with XCode 15.0.1
or later, build an app that will run on MacOS Sonoma and earlier
(and with luck, later).

This fixes issues #4384, #4385 and #4389.

New code copyright 2023 by John Dykstra.
This program is free software; you can redistribute it and/or modify
it under the terms of the GNU General Public License as published by
the Free Software Foundation; either version 2 of the License, or
(at your option) any later version.