Clean up loose ends in OS X support (#4443)
* Fix PYENV_VERSION undefined abort of a clean build

The GNOME jhbuildrc script incorrectly assumes that pyenv defines
the PYENV_VERSION environment variable.  This causes JHBuild to abort
at the start of a completely clean build.  The relevant code is unneeded
by QL, so delete it.

Fixes #4435.

* Remove obsolete FIXME comment.

This FIXME was previously resolved by one of the commits
squashed into 5e3989b.

* Eliminate duplication in bootstrap script

* Explain why Rust must be installed globally.

* Remove GStreamer "bad" and "ugly" plugins

Remove the "bad" and "ugly" GStreamer plugin projects from the MacOS
bundle.  These plugins are extremely unlikely to be added to a pipeline
by a user, and they needlessly increase the size of the app.

* Update dependencies on package gtk-doc

* Update the README in dev-utils/osx_bundle.

* Fix shellcheck warning