Properly handle control-click on MacOS as a synonym for right-click.
During builds for MacOS, patch gdk_event_triggers_context_menu()
in gtk+ to recognize control-click as a way to activate pulldown menus.
This is documented gtk+ behavior but was broken some time before
3.24.38 (see https://gitlab.gnome.org/GNOME/gtk/-/issues/6724).
This patch is their commit 4b5d503c, backported to gtk+ 3.

Update QL to use the GTK Python binding for this API.

    Fixes #4640.