Add SSL support to bundled Python for use by plugins.
Correct build of bundled Python interpreter so that it includes the
SSL module.  A bug in the 3.11.1 Python release (and others)
caused the "--with-openssl" configure parameter to have no effect
when the SSL module was built through setup.py.  Change the module
build specification so that it is built through Modules/Setup,
which correctly implements the relevant configure parameter.

Fixes #4648.